Downgrade to 6.15 + Add costume changes

+++ java/src/main/java/org/rocksdb/CompactRangeOptions.java
@@ -15,14 +15,13 @@ public class CompactRangeOptions extends RocksObject {
   private final static byte VALUE_kIfHaveCompactionFilter = 1;
   private final static byte VALUE_kForce = 2;
 
-  // For level based compaction, we can configure if we want to skip/force bottommost level
-  // compaction. The order of this enum MUST follow the C++ layer. See BottommostLevelCompaction in
-  // db/options.h
+  // For level based compaction, we can configure if we want to skip/force bottommost level compaction.
+  // The order of this neum MUST follow the C++ layer. See BottommostLevelCompaction in db/options.h
   public enum BottommostLevelCompaction {
     /**
      * Skip bottommost level compaction
      */
-    kSkip(VALUE_kSkip),
+    kSkip((byte)VALUE_kSkip),
     /**
      * Only compact bottommost level if there is a compaction filter. This is the default option
      */
