+++ util/crc32c_ppc.h
@@ -7,15 +7,12 @@
 
 #pragma once
 
-#include <cstddef>
-#include <cstdint>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 extern uint32_t crc32c_ppc(uint32_t crc, unsigned char const *buffer,
-                           size_t len);
+                           unsigned len);
 
 #ifdef __cplusplus
 }
